+++ core/src/com/superduckinvaders/game/entity/item/Upgrade.java
@@ -14,7 +14,7 @@ public class Upgrade extends Item {
     private Player.Upgrade upgrade;
 
 
-    public Upgrade(Round parent, double x, double y, Player.Upgrade upgrade) {
+    public Upgrade(Round parent, float x, float y, Player.Upgrade upgrade) {
         super(parent, x, y, Player.Upgrade.getTextureForUpgrade(upgrade));
 
         this.upgrade = upgrade;
@@ -23,10 +23,11 @@ public class Upgrade extends Item {
     @Override
     public void update(float delta) {
         Player player = parent.getPlayer();
-
+        /*
         if (this.intersects(player.getX(), player.getY(), player.getWidth(), player.getHeight())) {
             player.setUpgrade(upgrade);
             removed = true;
         }
+        */
     }
 }
